+++ gpt_core/src/error.rs
@@ -10,13 +10,4 @@ pub enum Error {
     Json { source: serde_json::Error },
     #[snafu(display("Reqwest error: {}", source), context(false))]
     Reqwest { source: reqwest::Error },
-    #[cfg_attr(
-        not(feature = "mock"),
-        snafu(display("Keyring error: {}", source), context(false))
-    )]
-    #[cfg_attr(
-        feature = "mock",
-        snafu(display("**Mock** keyring error: {}", source), context(false))
-    )]
-    Keyring { source: keyring::error::Error },
 }