+++ gpt_ui/src/cli/config.rs
@@ -1,5 +1,6 @@
 use super::dialog::{input, password, select};
 use super::error::Result;
+use encrypt_config::PersistSource;
 use gpt_core::config::Config;
 
 pub(crate) fn config() -> Result<()> {
@@ -19,37 +20,37 @@ pub(crate) fn config() -> Result<()> {
 /// Otherwise, the API key will be masked.
 fn display() -> Result<()> {
     #[cfg(feature = "mock")]
-    let config = Config::read()?;
+    let config = Config::load().unwrap_or_default();
     #[cfg(not(feature = "mock"))]
-    let config = Config::read_masked()?;
+    let config = Config::load().unwrap_or_default();
     println!("{}", config);
     Ok(())
 }
 
 fn modify() -> Result<()> {
-    let mut config = Config::read()?;
+    let mut config = Config::load().unwrap_or_default();
     let mut changed = false;
     match input(
         "Endpoint? [Empty to unchange]\n",
-        config.endpoint(),
+        &config.endpoint,
         #[cfg(feature = "cliclack")]
         false,
     ) {
         Ok(content) if !content.is_empty() => {
-            config.set_endpoint(content);
+            config.endpoint = content.parse().expect("Invalid URL");
             changed = true;
         }
         _ => {}
     }
     match password("API Key? [Empty to unchange]\n") {
         Ok(content) if !content.is_empty() => {
-            config.set_api_key(content);
+            config.api_key = content;
             changed = true;
         }
         _ => {}
     }
     if changed {
-        config.save()?;
+        config.store()?;
     }
     display()
 }